Add QSAutomation.[Status] table.  The contents are static and not used in code, but is a convenient reference for those using the tool.
Split StatusID 0 into two statuses:  0 (manually pinned by DBA) and 40 (best plan found)

High variation check now only skips status 0 and 30, and considers all other statuses.

Lowered the Default Mono Plan Performance Threshold to 2000ms

Lowered the Mono Plan execution count minimum to 5.  Prior to this it was set to 100.  This issue with this high number is that queries with multiple plans but only one plan with >100 executions appeared to be mono-plan when they really were not.  Lowering the threshold helps with this issue, and allowing mono-plans (StatusID 40) to be evaluated by the High Variation Check ensures that mono-plans are re-evaluated regularly.

Sending an email when the Query Store is reset is controlled by the log level now instead of on an incremental basis.

When preparing to unpin a plan to allow new ones to form, check first to make sure the plan is still pinned (otherwise you will get an error)